fix(fiatconnect): update mobile money to match fc spec (#3329)
### Description

When adding a Mobile Money account, the wallet currently provides a
dropdown showing a set of supported network operators by default. The
fiatconnect specification was
[updated](https://github.com/fiatconnect/specification/pull/88) such
that we should no longer show this dropdown by default, and instead we
should allow users to enter the name of their operator using their
keyboard. This PR implements this functionality. This PR also updates
the both the `fiatconnect-sdk` and `fiatconnect-types` packages to the
latest version which no longer features the `SupportedOperatorEnum`


![UpdatedMobileMoneyOperator](https://user-images.githubusercontent.com/20508929/210897290-4b2335aa-30dc-4c04-84da-71fd164d4023.png)


### Test plan

Manually verified adding a mobile money account behaves as expected.

### Related issues

- ACT-581

### Backwards compatibility

Yes.