fix: android restore wallet flow navigation stall (#3332)
### Description

There are 2 problems this PR solves:

#1: on android, sometimes the recovery phrase text box does not appear
(i find this reproducible if i land on this screen with a recovery
phrase already on my clipboard). Video of the issue:


https://user-images.githubusercontent.com/20150449/211043417-0ded2018-6cec-45b8-a77a-7d87f5598af3.mp4

This one is annoying, I am fairly sure it has something to do with the
Layout animation of the "paste" button. The value of the clipboard
content is always an empty string to start with (since fetching the
clipboard content is async) so the `ClipboardAwarePasteButton` component
always needs to do the layout animation twice as `shouldShow` changes
from `false` initially to `true` when the clipboard content is picked
up. The issue does not exist if this `shouldShow` prop doesn't change
value between renders, i.e. if there is no recovery phrase on the
clipboard). I couldn't find a way to fix this, so i just disabled this
animation since it doesn't seem like a critical feature....


#2: the navigation seems to be broken after the recovery phrase is
entered. on force close then restart, the app does function normally.
video below:


https://user-images.githubusercontent.com/20150449/211043522-4aee3914-f136-46ff-ae6e-4ed78acfc6ac.mp4

Since we've disabled the swipe to navigate back on iOS and we have
control over the hardward back button on android, there's no more need
to adjust the navigation stack. The change in this PR removes the
navigation stack logic and makes the back button on android (and the
cancel button on the header) go to the welcome screen. Video after the
changes in this PR:


https://user-images.githubusercontent.com/20150449/211043464-107125df-ce3d-42bb-9d6e-b16d2c32b495.mp4


### Test plan

Test on iOS and Android:
- restore flow
- navigating back on restore wallet screen
- create flow

### Related issues

- Fixes RET-562

### Backwards compatibility

Y

Co-authored-by: Tom McGuire <Mcgtom10@gmail.com>
Co-authored-by: Joe Bergeron <jbergero@alum.mit.edu>
Co-authored-by: Silas Boyd-Wickizer <silas@valora.xyz>
Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
Co-authored-by: valora-bot-crowdin <80050969+valora-bot-crowdin@users.noreply.github.com>
Co-authored-by: Valora Bot <89419329+valora-bot@users.noreply.github.com>
Co-authored-by: valora-bot <valorabot@valoraapp.com>