feat(fiatconnect): cache transfers & quotes by tx hash (#3401)
### Description

Currently, there is no mechanism to store a successful FiatConnect
transaction. This is needed because in ACT-578, we plan to populate
FiatConnect transfers in the transaction feed with greater details about
the transaction. This PR adds a new `cachedTransfers` array to the redux
state, and updates the FiatConnect saga to append successful
transactions to it.

### Test plan

Added new unit tests

### Related issues

- ACT-579

### Backwards compatibility

Backwards compatible