fix: low allowance swap issues (#3421)
### Description

We are calling 0x api twice, one to get the details of the swaps (the
amounts and prices) and another to confirm everything is ready to
execute the swap. This increases how long the swaps take to complete,
two fetches instead of one and produces errors with the guaranteed price
when the buy amount is set. To simplify we are now using the transaction
object returned from the `/approveSwap` endpoint instead of subsequently
calling the `/executeSwap` endpoint and using the swap transaction
present in that response. This is more inline with the [0x API
docs](https://docs.0x.org/0x-api-swap/guides/swap-tokens-with-0x-api)
and will make future swap router implementations easier.

### Test plan

Tested locally on Android and saga tests updated.

### Related issues

- Fixes RET-572

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->
Yes