chore: turn offline banner alert into non blocking toast (#3424)
### Description

Aside from this blocking banner, we don't currently have another way to
show users that their balance is not up to date. Until we have a way to
let users know that they are back online, i think we have no choice but
to leave some kind of messaging on the screen. I see instagram uses a
toast for connectivity issues, which i quite like so thought i'd try it
here.

I like it because it does not forfeit any existing functionality. Users
are still able to trigger a refresh via the toast, and they can still
navigate the app. The toast persists until the problem goes away, so
that users can know that their balance is up to date again. I also like
that it's a relatively unintrusive change + we can reuse an existing
component.

There is some extra state management to preserve the toast dismiss
animation. Since there is only one alert active at a time, there could
be a scenario where there is an existing toast -> a new error banner is
triggered. Without some extra state management, the toast would display
the new error banner message while animating off the screen. This is a
very small issue of course but it was also simple to fix with an extra
state variable. Example with slow animation:


https://user-images.githubusercontent.com/20150449/216957689-5ee1c979-5422-479d-a5ea-6e330d572fb3.mp4



### Test plan



https://user-images.githubusercontent.com/20150449/216770235-3b2859d2-57a3-4a7a-8fbf-2046b33710b9.mp4



### Related issues

- Fixes RET-597

### Backwards compatibility

Y