feat: dapps bottom sheet dynamic height based on content size (#3412)
### Description

We should have the bottom sheet height respond to the content size
according to the new design. There's no out of the box way to do it yet
using the bottom sheet navigation lib (but there is an[ open
request](https://github.com/th3rdwave/react-navigation-bottom-sheet/issues/7)
to support it, with an example of how someone else did it). The issue
has been open for a while but there's a chance that it could be
supported soon, so opted for a patch for now so that we can still use
the `useBottomSheetDynamicSnapPoints` hook.

Not addressed in this PR:
- handling displays for large text (probably need to add scrolling)

### Test plan

With WalletConnect:


https://user-images.githubusercontent.com/20150449/215788229-e2eaf2fd-ec68-4b6f-8736-7ec72522628b.mp4


With DappKit:


https://user-images.githubusercontent.com/20150449/215788377-dca31a5d-8d54-42f2-91af-4c6204ba89f8.mp4


### Related issues

- Fixes RET-551

### Backwards compatibility

Y

---------

Co-authored-by: valora-bot-crowdin <80050969+valora-bot-crowdin@users.noreply.github.com>
Co-authored-by: Valora Bot <89419329+valora-bot@users.noreply.github.com>
Co-authored-by: Jean Regisser <jean.regisser@gmail.com>