feat: display FiatConnect transfers differently in transaction feed (#3407)
### Description

Currently, the wallet display FiatConnect transfers in the same style as
any other transfer in the transaction feed. This PR adds new
functionality to check if a transfer is a FiatConnect transfer, and if
so populates the transfer feed item so that it matches the style shown
in this [figma (the Withdraw cUSD
row)](https://www.figma.com/file/bwIY7peeykI8K4wTrXs4iN/Fiat-Connect?node-id=4264%3A49583&t=hvdNzAVgoJ4zkEkv-0).



![fcTransactionFeedItem](https://user-images.githubusercontent.com/20508929/215867211-a5b4b00c-00f3-4524-b1be-ea785e9027f6.png)


### Test plan

Added unit test.

### Related issues

- ACT-578

### Backwards compatibility

Backwards Compatible.