chore: upgrade react-native-gesture-handler to v2 (#3440)
### Description

We mostly don't need to use this dependency directly. I think that we
were using some components from there because of auto import
suggestions, in this PR i updated the touchables to be used from either
the Touchable component or from react native directly. Also simplified
some styles along the way.

This is a good one to upgrade because it removes the dependency on
RNGestureHandlerEnabledRootView which was causing conflicts while
upgrading react native.

### Test plan

Tested all the changes manually. Relying on CI to check the rest.

### Related issues

- Fixes RET-587

### Backwards compatibility

Y