refactor(onboarding): consolidate onboarding flow logic. (#3505)
### Description

This solves a few problems with the current onboarding flow logic:
1. The logic for which screen to go to next in onboarding is currently
spread out throughout the different onboarding screens. This makes it
difficult to see in-code how the onboarding flow actually works with its
different branching options. With this refactor all that logic would be
in one place
2. The logic for the 'Step X of Y' is currently totally separate from
the logic of which screens come after one another in onboarding. This
duplication of logic is confusing and messy. With this refactor we have
a traversable graph that can be used to calculate the steps without
separate logic that must change every time a screen is added.

### Test plan

1. More unit test coverage than there was before for the onboarding flow
2. Unit tests of the Onboarding directed graph ensuring that counters
always increment and the graphs always end up at Home.

### Related issues

https://linear.app/valora/issue/ACT-597/refactor-onboarding-screen-logic

### Backwards compatibility

yes

---------

Co-authored-by: jh2oman <jacobwaterman@MacBook-Pro.local>
Co-authored-by: Charlie Andrews-Jubelt <cajubelt@users.noreply.github.com>