fix(wc): prevent unhandled errors from cancelling all WC action handlers (#3527)
### Description

This PR fixes the issue observed where any unhandled error in a WC
action handler would cancel all other listeners for WC actions. The only
way to recover was to restart the app.

It uses the `safely` helper added in #3526.

Note: there are other places in the codebase suffering from the same
issue. We can open additional PRs for them.

### Test plan

- Added unit test

### Related issues

- Fixes RET-654

### Backwards compatibility

Yes