chore: ensure supercharge v2 first session data refresh (#3548)
### Description

Supercharge v2 is controlled by a remote config variable. When the app
is launched, we fetch supercharge rewards and remote config at around
the same time resulting in a race condition. It is very probable that
the user will fire a request to fetch supercharge v1 rewards before the
updated remote config value is received. This means that the user could
navigate to the supercharge screen and try to claim their v1 rewards,
which will always fail because the superchargeV2 remote config has
updated by this time.

### Test plan

There are two ways the user can visit the supercharge screen:
1. open the app -> navigate to supercharge screen
2. open the app through a push notification -> land directly on the
supercharge screen

In both cases, it is most probable that the request to supercharge v1
will be fired before receiving the updated remote config value. These
are the ways I think the race condition could play out:
1. the updated remote config variable comes back before the supercharge
v1 response: we will prevent the v1 response from being saved if
supercharge v2 is enabled.
2. if the supercharge v1 response comes back first and is saved, when
the updated remote config value is received then we will clear the saved
rewards and fetch supercharge v2 rewards.


### Related issues

- Fixes RET-656

### Backwards compatibility

Y