chore(test): update Trans mock to include translation key and params (#3562)
### Description

The existing mocks for the `Trans` component only renders the children.
In unit tests, it is therefore not possible to make an assertion like
"expect this translation key to have been rendered". I think that it's
important to be able to verify the translation key / params as well as
the child components. This PR changes the mocks to include the
translation key and param, as we do in the mocked `t` function.

### Test plan

n/a

### Related issues

n/a

### Backwards compatibility

Y