feat: sprinkle haptic feedback in a handful of places (#3563)
### Description

This PR adds the plumbing to do haptic feedback, and adds it to some key
actions. I added a new lib because the `Vibration` that comes with react
native is very limited. On iOS, the vibration is fixed at 400ms which
was way to aggressive.

This PR:
- add haptic feedback lib
- add a haptic feedback helper file - so that we can give the vibrations
semantic meanings and change them easily in one place
- add haptic feedback to in app notifications, primary buttons, copy
action, dapp favourite, settings toggles

### Test plan

Manual

### Related issues

- Fixes RET-651

### Backwards compatibility

Y

---------

Co-authored-by: Tom McGuire <Mcgtom10@gmail.com>