feat(fiatExchange): Don't show add funds when pressing withdraw home action (#3610)
### Description

Fixes ACT-705. Adds a new version of the add/withdraw screen, without
the "add" option, and navigates to it from the home carousel.

There were two "issues" with this.
1. Components used for Drawer navigator screens cannot accept props;
Typescript will yell at you.
2. We will want to be able to _not_ have the Add/Withdraw section appear
in the drawer nav, but make it navigable from the home actions section.
In order to do this, we must _not_ register the Add/Withdraw screen with
the Drawer navigator (so it doesn't appear in the side nav); removing
the screen from the Drawer navigator and _not_ adding it to our regular
navigator makes it un-navigable, so we must somehow register it with our
normal navigator.

In order to do this, the `FiatExchange` screen was component-ized and
parameterized. Two wrappers were created, neither accepting props; one
to be used by the drawer navigator (with behavior unchanged) and a new
which removes the add funds section and drawer nav header.

### Test plan

Unit and manual tested.


https://user-images.githubusercontent.com/569401/229833998-b8346a89-445f-4b4c-a6f8-dc70cb4b74a1.mp4

### Related issues

- Fixes ACT-705.

### Backwards compatibility

Yes.