fix: dismiss keyboard on bottom sheet open (#3625)
### Description

Noticed when working on #3599. Currently if the keyboard is visible the
when `openSheet` is called the bottom sheet will display behind the
keyboard. If there is no keyboard visible on the screen, then the
`Keyboard.dismiss()` method will not have any effect.

#### Added to
- `openSheet` from `src/dappsExplorer/useDappInfoBottomSheet.tsx`
- `src/components/BottomSheet.tsx`

#### Before

https://user-images.githubusercontent.com/26950305/230246866-57bd2246-cb55-4d79-9672-6712e071fb9b.mov

#### After

https://user-images.githubusercontent.com/26950305/230247068-85bc3ada-f6dc-4302-9537-e0e30c3c2642.mov

### Test plan

Tested locally on iOS and Android

### Related issues

N/A

### Backwards compatibility

Yes

---------

Co-authored-by: Jacob Waterman <jacobrwaterman@gmail.com>