feat(home-actions): Show back button on screens reached from home actions (#3643)
### Description

Part 2 of ACT-704. Currently, the screens reached from the home actions
can display the hamburger menu or the 'X' icon in the top left of the
screen. This PR replaces these navigation methods with a back button
only when the screens are accessed from the home actions. When reaching
the screens via any other methods, the existing navigation option will
still be shown. Note that navigation for the swap screen is handled in
#3623.


https://user-images.githubusercontent.com/20508929/232064472-422cc988-c514-406b-b0ee-06d94d17e4c5.mp4


### Test plan

Added unit tests and updated snapshots.

### Related issues

- ACT-704

### Backwards compatibility

Yes.