chore: remove duplicate openssl pod (#3695)
### Description

I forgot that the upgrade RN 67 branch had automerge enabled, and it was
merged before I could tidy this up.

There was some change in RN 67 where OpenSSL was no longer part of the
release build. This causes a crash on app start. This
[article](https://medium.com/@devsalaar/dyld-library-not-loaded-error-on-ios-react-native-fix-guide-bb0b6a91e85a)
was helpful for debugging why this happens - it seems that the
use_flipper script adds the OpenSSL pod only for debug configurations.
Maybe there is a better solution here but it seems like we just need to
manually add OpenSSL.

### Test plan

n/a

### Related issues

- Fixes RET-683

### Backwards compatibility

y