feat: add sticky segmented control in assets screen (#3762)
### Description

This PR:
- add a new segmented control component (the existing one used on the QR
screen is tightly coupled to the react navigation tabs implementation, i
couldn't figure out how to easily extend it. it also looks different...)
- add animations for a sticky segmented control, where part of the list
header is not sticky
- on tap of the segmented control, update a state value. displaying the
positions isn't included in this PR.
- no unit tests yet because the segmented control doesn't do anything

Note: i need to change the Flatlist to a Sectionlist for the positions
display, but Sectionlist doesn't allow a sticky header (only the section
headers can be sticky). That is why we are not using a
ListHeaderComponent here anymore.

### Test plan

iOS:


https://github.com/valora-inc/wallet/assets/20150449/84c2d455-bd5f-471f-9666-86138fa7667c

Android - the screen contrast isn't good here but there is a slight
shadow when the header becomes sticky:


https://github.com/valora-inc/wallet/assets/20150449/3928cc89-b96d-49e9-8cf9-11e8d7ff3fb9

Android, without the segmented control:


https://github.com/valora-inc/wallet/assets/20150449/823113f7-ac43-49bd-817d-2c4d8b5333e8



### Related issues

- Fixes RET-734

### Backwards compatibility

Y